Translated multi-threading works (though with bookkeeping errors)
Updates required to ensure that alphabets were correct, and that
blocks were correctly allocated to pass DNA sequence into
translation code, and to receive amino acid sequences out.

The set and scores of matches are in agreement in serial and
threaded versions, but there appear to be erros in bookkeeping
that are leading to mistakes in adjusting E-values.