Fix counting/naming of ORFs. Other small cleanup/output change.
(1) Modify counting of ORFs going through the pipeline to account
for duplicates arising from the fact that we work with overlapping
windows of dna sequence (so we don't miss orfs at the boundary of
a window). Also adjusts size of overlap and the naming of ORFs so
they are consistently named and counted the same regardless of # of
threads used.

(2) Modify p7_pli_Statistics to write out the number of "orf(s)"
instead of the number of "sequence(s)"

(3) Remove unused range-restriction feature of hmmsearcht
(--restrictdb_n, n_targetseqs)

(4) Removed a line added to p7_alidisplay.c back in commit 780b12a98
that sets ad->ntseq = NULL (to remove problems with serialization.
This line causes translated alignments to no longer include the
nucleotide sequence. I believe this is an unnecessary line, because
of the existence of an entry a few lines earlier (from commit
33fb702b6) that sets ntseq = NULL for all but translated search ...
but will be working on the daemon next, so will soon validate against
utest alidisplay_SampleFake_ntseq(), etc.

hmmsearcht and hmmscant are now generally working as expected in my
tests. hmmsearcht still needs adjustment to get pli_statistics() to
print orf count instead of sequence count; hmmscan still fails 'make check'