Further revision to the fix for issue #176: the test itself requires ~10G RAM.
p7_gmx_utest still failed under Travis-CI. I got it to fail on gcc13
in GCC compile farm in France. gcc13 has limited RAM (4GB), and indeed
the problem traced simply to a failed malloc().

The unit test for #176 requires 8.6GRAM in a single malloc() to
exercise the bug. Added code to the unit test to check that a 10G
allocation succeeds before trying to do the test.