Make sure that model parameterization doesn't depend on RF presence/absence.
esl_msaweight_PB() uses RF consensus column annotation if present, so
the same MSA with versus without RF annotation may get different
consensus cols. Easel sees this as a feature; MSAs with vs. without RF
annotation are different inputs, and weighting is faster with an RF.
However, in HMMER and Infernal we don't want model parameterization to
depend on presence/absence of RF annotation; in default
hmmbuild/cmbuild, we want to ignore any RF annotation, and with
--hand, we want to use it. (And in jackhmmer iterations, which are
always --hand, we want to use it too.)

Eric modified Infernal to use a custom esl_msaweight_PB_adv()
configuration that sets ignore_rf to FALSE when --hand is set, else to
TRUE. This commit introduces the corresponding changes to H3.

This change required updates in regression tests in
i18-nhmmer-generic.pl. This test uses testsuite/3box.sto, which has an
RF line; default parameterization of this model is affected by the
change. Also updated 3box.hmm to the new default construction.

See EddyRivasLab/hmmer#180 for discussion with Eric.
[SRE H8/27]