Force trace to be Reuse()'d after p7_Decoding() numeric overflow
In p7_domaindef.c, following a call to p7_Decoding(), line 897 was:
   if (status == eslERANGE) return eslFAIL;
The comment says this is the result of repetitive garbage [J3/119-212],
which is certainly the case that has caused it to fire in my hands.

The eslFAIL return value is handled gracefully in the calling function,
but since it returns immediately, the normal end-of-function cleanup
Wasn't happening. That cleanup is basically just the line:
   p7_trace_Reuse(ddef->tr);

The result is that the next call to p7_OATrace() failed with:
   Fatal exception (source file optacc.c, line 233):
   trace not empty; needs to be Reuse()'d?
   Abort trap: 6

The solution (committed here) is to force the trace to be reused
before returning eslFAIL:
   if (status == eslERANGE) {
          p7_trace_Reuse(ddef->tr);
          return eslFAIL;
   }