Fix for iss#171: guessing query file format
If nhmmer is called without guidance on the input type, it tries to guess
whether the query is an HMM, MSA, or unaligned file of seqs. It previously
did this by trying to open the file first as an HMM, then as an MSA, and
finally as a sequence file - if it succeeded at any point, it would decide
this was the correct input type. This over-eager guessing strategy could
lead to unexpected results:
- when reading a sequence file with a single sequnce, nhmmer would treat
the input as a one-sequence MSA; this is not intuitive.
- when reading a sequence file with many sequences all of the same length,
nhmmer would guess that this was an MSA, and smash them into a single query,
computing one model for the "MSA"; this is just plain wrong

This commit reimplements the format guessing framework, still seeking to
get the guess right, but insisting that the user provide guidance if the
choice isn't completely clear. The strategy is:

- If the user has provided guidance w/ --qhmm, --qmsa, or --qseq, do as
instructed.  (note: --qseq replaces the previous overly-restrictive
--qfasta)

- Otherwise, try to open the file as an HMM. This either works or doesn't,
with no risk. (note: if this fails and the file is piped, it can't be
rewound, so give a "must specify format" message)
- If the file is rewindable, check to see if it's obviously an MSA file
or obviously a sequence file. If not obvious, we'll force the user to
tell us. That looks like this:
   - Try to open it as an MSA file
      - if ok (i.e. it opens and passes the MSA check, including that
        all sequences are the same length)
         - if it's a FASTA or A2M format, it still might be a sequence
           file
              - if the "MSA" is a single sequence, then rewind and call it
                a sequence input.  Otherwise give "must specify" message
         - otherwise, it's an MSA;  proceed accordingly
      - if not ok (i.e. it's not an MSA file)
         - if it's anything, it must be a sequence file, proceed accordingly

(This commit is not ready for prime time yet. Still need to fix handling
of the --qformat flag, and bring documentation up to speed)