Revised fm_general.c error handling, messaging
In the original implementation, the fm_general.c code for
reading FM index data would exit (via a call to esl_fatal)
in the case of errors. When nhmmer attempted to autodetect
target file format for a bad file (one that was neither a
recognized sequence file nor an FM file), the logic would
fall into the FM-reading code, then run into a reading error
and die by esl_fatal with an unhelpful message that
complained of FM-specific errors, when the real error was
in autodetection. This is issue #195.

In this fix, fm_general.c errors in reading metadata now
return an error code, and the calling function can
gracefully report that autodetection failed (if that's
the case)