Fix handling of discordant user input format
Fix error in esl_sqfile_Open() call that allowed format-guesser to
override the command-line mandated format

Also smaller changes:
- Remove superfluous esl_msafile_EncodeFormat() call
- Previously, if input is predicted to be a2m, would complain that
it might be single seq or multiple alignment format. This was wrong:
if the file is labeled as a2m, it means it was rejected by the
autodetection code as being fasta, and accepted as the (always MSA)
a2m format. The a2m call is now respected.