Enable forcing msa query to be read as single seqs
Sean pointed out that phmmer allows msa formats to be read as single
sequence sources, and suggested providing that nhmmer include an
override flag to do that. I've added it:  --qsingle_seqs

The logic is pretty simple: if --qsingle_seqs is enabled, then open
the file with esl_sqfile_Open() instead of esl_msafile_Open(). This
is true whether the specific format was specified with --qformat or
is being autodetected.

Still need to update documentation to track with this change.