Handle afa w/ gaps, when user requests single seq
After adding the --qsingle_seqs flag, there's a remaining situation that
might be confusing to a user: when the query is in aligned fasta format,
containing gap characters. The user might (I think reasonably) expect
nhmmer to just ignore the gap characters.

Run without flags, nhmmer complains that it can't autodetect (as it should):
% nhmmer seqs6-gaps-samelen.afa seq1.fasta
Error: Query file type could be either aligned or unaligned; please specify (--qformat [afa|fasta])

Run in what I'd call the "usual" way when using an afa query, it works:
% nhmmer --qformat afa  seqs6-gaps-samelen.afa seq1.fasta | grep -c "Query:"
1

The problem would arise when the user indicated that the gap-filled afa
file was an unaligned fasta file, either directly (--qformat fasta) or
indirectly (--qsingle_seqs).  The result was an error
    Line 2: illegal character -
caused by the fact that the single-sequence fasta reader dislikes
gap character.

The solution is to catch the eslEFORMAT error induced by finding a
gap character in a fasta sequence, notice that we think it's a fasta
file (which, due to logic flow, would only happen in the case of one
ot the flags mentioned above), then try it again as an afa file
(which handles the gap characters gracefully). That approach is
implemented here.