nhmmer fails if given AFA, but user asserts FASTA
Sean has indicated that it's correct behavior for nhmmer to
complain with an error if the user tries to assert that an AFA
alignment file (with gaps) is an unaligned FASTA file.  I've made
that change here. If the user adds --qsingle_seqs, that instruction
will dominate, forcing nhmmer to treat all sequences in the file as
individual sequences, whether the user specified a qformat of AFA
or FASTA or nothing.  (Other formats continue to work as expected)

Note: A change to documentation is still pending.