Improving several nhmmer query-reading errors
If the user supplied an asserted format, this format is now
added to errors that suggest there's something wrong with the
format. As an example: if the user supplied a fasta-formatted
file, but adds the flag '--qformat stockholm', nhmmer doesn't
have time to confirm that this format is actually fasta -
it dies while trying to guess the alphabet based on a wrong
format assertion. Various error messages now add text that
indicates what format it's been told the file is in, in hopes
that this will help the user diagnose the incorrect format
assertion.