p7_Pipeline() will now throw an exception on target seqs > 100K length
Elena reported a bug where `hmmsearch <DNA model> <genome db>` gives
inf scores, because it happily aligns the model to entire
chromosome-length target sequences. She really meant to use nhmmer,
but nothing warns her. The numerics of the comparison engine assume a
max target seq length of 100K but this is enforced nowhere.

Because this is a normal (user) error, and p7_Pipeline() is not a
top-level application (and indeed, can be called within MPI or
threaded applications) code style to return a normal error (w/
optional error msg) to the caller. Appropriate error is <eslEINVAL>
(invalid argument), and we have the error message space in
<pli->errmsg>.

However, we're already using <eslEINVAL> for a different error
(hmmscan using GA/TC/NC thresholds but model doesn't have any set),
and we're only checking for that error where it can occur (in
hmmscan). In general we aren't checking return status on p7_Pipeline()
calls, so it's not straightforward to drop a new type of error in
without rethinking MPI, threads, and other calls.

So instead, I handle it as an eslETYPE exception, and I added a note
that this is a workaround that could be improved.