Fix for bg reparameterization failure w/ repeats
This is a fix for https://github.com/EddyRivasLab/hmmer/issues/198. The
change is simple: in rescore_isolated_domain(), if p7_Decoding() returns
eslERANGE (assumed to mean that the target appears to be repetitive
garbage), then the bg model is returned to its original form via a call
to reparameterize_model() before rescore_isolated_domain() returns
an eslFAIL status. Previously, it just returned the status without
recovering bg.