Fixes hmmpgmd (and iss159) unit test failures when built with --enable-mpi
On OS/X with OpenMPI, when built with --enable-mpi, three unit tests
failed: i19-hmmpgmd-ga, i22-hmmpgmd-shard-ga, and
iss159-nhmmer-overlap.

The iss159 test fails because of a side effect of either hmmpgmd test
failure: those failed tests leave press'ed HMM database files, and
when iss159 creates a new HMM with the same TMPPFX and tries to open
it, the (protein) hmm db files are opened instead. We currently don't
have any good checking for the case where a new .hmm file has
overwritten an old one that was hmmpress'ed, creating a mismatch
between the .hmm file and the .hmm.h3? index files.

The hmmpgmd unit test failures are in how hmmc2 handles a normal
shutdown. It sends a !shutdown command; the master processes it and
shuts down the worker(s) and itself; hmmc2 then attempts to do a
readn() call which should fail, because the master has shut down. The
errno that gets set when readn() fails, though, appears to vary
depending on OS and compiler, in a way that I don't understand, and
that we should look at harder someday. Farrar was checking for
ECONNRESET; we already have a Mar 2019 commit (3b0fe98) from Nick
adding a check for 0 to fix a bug here; my fix is to also check for an
ESRCH (no such process) error and treat that as a normal shutdown too.
So now we're considering (ECONNRESET | ESRCH | 0) as normal shutdown
of hmmc2.