Fix for nhmmer memory error - issue 213
This commit is a fix for https://github.com/EddyRivasLab/hmmer/issues/213.
The error was caused by incorrect sizes for some of the sequence elements
in the data structure underlying the FM-index-based binary representation
of the sequence database created with makehmmerdb. The error was a simple
matter of failing to initialize the value to zero before accumulating
length while reading the sequence (i.e. lengths could be effectively
random under certain memory access conditions).

Though the original issue only found an error when nhmmer search was
performed on a database produce by running makehmmerdb with the
--block_size flag, the error also could affect databases produced without
any makehmmerdb flag.