Fixes hmmpgmd test failures on OS/X Big Sur with --enable-mpi
When hmmc2 is shutting down the server and tries it again (to be sure
it's shut down), it gets an errno code back from the failed readn().
Exactly what this code is, isn't well standardized. hmmc2 has a little
(growing) list of errno codes it accepts as normal. For whatever
reason, on OS/X Big Sur (but not Catalina), with --enable-mpi (but not
otherwise), the errno is 1, EPERM Operation not permitted.

The fix is to add EPERM to the little list.

Additionally, the iss159-nhmmer-overlap test fails if an earlier
hmmpgmd test failed and left tmpfiles around. This is because the
hmmpgmd <tmppfx>.hmm.* index files conflict with the <tmppfx>.hmm that
the iss159 test creates.

The fix is to have the iss159 script clean up any tmppfx.hmm.* index
files before it creates its HMM.

[xref SRE:H10/106]