Fix to seg fault in nhmmer fm search (#257)
This fix is intended to fix the seg faults documented in iss #257
(and previously in discussion for PR #253): seg faults that occur
under some conditions when using nhmmer to search the digitally
formatted target database produced using makehmmerdb.

The errors were caused by the final line in the FM_backtrackSeed()
function in fm_ssv.c:
  return len + (j==fmf->term_loc ? 0 : fmf->SA[j / fm_cfg->meta->freq_SA)) ;

That function plays a role in identifying the location in the target
sequence at which a particular sequence occurs. The process involves
tracking back through a down-sampled suffix array until a sampled
position is reached, then returning the number of steps (length) of
the backtrack. If the backtracking step ever runs into the suffix
array sentinel character ('$'), the back-track stops, and the returned
value (len) is expected to by adjusted by the value in the final entry
in the downsampled suffix array.

The problem:  (fmf->SA[j / fm_cfg->meta->freq_SA) is the number of
entries in the array ... so len was being adjusted by whatever value
is found in the memory location just beyond that entry. That value
can be ... anything, so the function was returning an effectively
random number, dependent on what the memory layout caused to be placed
at that n+1'th memory location.

Backtracing through the sentinel character is expected to occur
infrequently, but on a large enough search space, it's bound to
happen. The impact of this error would be unpredictable, and would
depend on what ~random value was stored in the accessed memory
location: it could cause an incorrect location to be identified
(possibly causing the loss of a hit), or the seg faults that were
sometimes observed.

The fix is easy - just reduce the index by 1:
       fmf->SA[ (j / fm_cfg->meta->freq_SA)-1 ]
This only impacts the case where the sentinel was reached, and is
the correct index to use.