Fix to seg fault in nhmmer fm search (#257)
This fix addresses the seg faults documented in iss #257 (and
previously in discussion for PR #253): seg faults that occur
under some conditions when using nhmmer to search the digitally
formatted target database produced using makehmmerdb. This is
the second commit addressing this error - the first was incorrect,
and since reverted.

The problem was actually just that the allocation of the sampled
suffix array (SA) was one too small. The math is:
- the full sequence is length n, so an unsampled SA has n entries
- sampling every kth entry, there will be floor(n/k) entries
- for consistency with other parts of the FM data structure,
  indexing of the SA starts at 1. Unfortunately, the allocation
  assumed base-0, i.e. it allocated floor(n/k) ints, and should
  have allocated 1+floor(n/k).
- This posed a problem when trying to access the final (n/k)th
  element in the sampled SA, as is done in the final line in the
  FM_backtrackSeed() function in fm_ssv.c:
    return len + (j==fmf->term_loc ? 0 : fmf->SA[j / fm_cfg->meta->freq_SA)) ;
  (here: n==j, and k==freq_SA). For the largest few values
  of j, (j/freq_SA) refers to the memory location just beyond
  the allocated sampled SA. That value can be ... anything, so
  the function was returning an effectively random number,
  dependent on what the memory layout caused to be placed at that
  memory location. This could cause either an erroneous result
  or a segfault (in this small number of cases).

The solution is to increase the number of allocated values in
the sampled SA by one. This needs to be done both in the fm_ssv
code that raised the error, and also in the code that produced
the stored FM index (in makehmmerdb.c).

Note: this means that the db produced by previous instances of
makehmmerdb are broken: the final value in the sampled SA was not
written to disk. Any attempt to re-use an old pre-computed binary
target will fail.

While making these changes, I also caught another allocation error
in makehmmerdb: the value of max_block_size was sometimes under-
allocated by one due to rounding down when allowing for the
read_window block size adjustment. Taking the ceiling of
(block_size*.05) overcomes this error.    ~line 536:
  max_block_size = FM_BLOCK_OVERLAP+block_size+1  + ceil(block_size*.05);

These changes were tested across a number of inputs, and validated
using address sanitizer.