Fix alimask out-of-range errors (iss #237) (#258)
* Fix alimask out-of-range errors (iss #237)

This commit addresses issue #237:

"Alimask crashes if asked to mask an alignment in a region that's
outside the range of the alignment."

I've added range checks on all four of --alirange, --modelrange,
--ali2model, and --model2ali. For all four:
- the smallest among all <from> positions must be >0
- for all ranges, <to> can not be less that <from>
- the largest among <to> positions is restricted as is described
in the next paragraphs.

For (--alirange, --ali2model): <to> can not exceed the alignment
length.

For (--modelrange, --model2ali):
- alimask works as before, computing the model that would be produced
  by hmmbuild (by numerous flags mirroring those from hmmbuild).
- for all ranges, <to> can not exceed the resulting model length.

The --ali2model required a bit of extra attention:
The purpose of this flag is to print, for each alignment range
<from>..<to>, a mapping of the range with two pairs i..j => m..n,
such that
- i<=j
- i is the first alignment position between <from> and <to>
(inclusive) that maps to a model position; that position is m
- j is the final alignment position between <from> and <to>
(inclusive) that maps to a model position; that position is n

There were two bugs in this code. The first allowed j to exceed
<to> if the <to> column did not map to a model position. This
is fixed. The second allowed a range-pair i..j => m..n to be
produced even if no alignment columns in the range <from>..<to>
map to a model position. I've revised the mask output to
indicate this lack of mapped positons with an output like:
          <from>..<to>       ->       -..-  (no map)

* Free allocated vars, fix valgrind leak errors

* Fix ali2model error for out-of-order ranges

When multiple --ali2model ranges are provided, and they
are out of order, the previous implementation would fail
to report the mappings for ranges with smaller indices
that appear after ranges with larger indices, e.g.

   --model2ali 14..15,7..9,37..47

alignment coordinates     model coordinates
         14..15       ->       14..15
          7..9        ->       -..-  (no map)
         37..40       ->       29..31

This is now resolved.

* Fix error typo

* Improve help messages, fix a couple typos

* Update appendmask argument

Previous commit included a change to a misspelled argument
(--apendmask -> --appendmask). These changes should have come along
with that change.