Fix FM error w/ ambiguity codes in consensus check
The FM-accelerated form of nhmmer includes an edge-case
test that allows below-threshold matches to remain
un-pruned if the target sequence seed is an exact match
for the matching part of the consensus query. The FM
implementation does not expect ambiguity codes in the
query models, which seems like a reasonable limitation
since there's no specific character to be exactly-matched
to.

When working with models built by hmmbuild, this is not a
problem: all positions have some letter from [ACGT] as the
consensus, even if the input sequence contained ambiguity
characters.  But when nhmmer works with single sequence
queries, it builds the model with p7_SingleBuilder(),
which calls p7_hmm_SetConsensus(hmm, sq), which causes the
consensus to be set to match the letter in sq ... which
might be an ambiguity character.

This poses a problem in fm_ssv.c, when the reverse complement
consensus is being computed:
   cons_c = fm_cfg->meta->compl_alph[consensus[k]]
causes cons_c to be set as undefined (-1). Bad memory
accesses happen downstream.

The fix: when reading the consensus into the fm_ssv data
structure, if an ambiguity character is seen, replace it
with a random letter (fixed seed) from [ACGT]. This will
produce a "consensus" that will be accepted in downstream
fm-related methods. A run of such positions is highly
unlikely to survive the seed-finding stage, and if it does,
it'll simply be filtered out at the next SSV-scoring stage.

(Note: this consensus sequence check is just a safety
mechanism to keep a few low-scoring but attractive seeds
around; the meat of the FM filter works just fine with
ambiguity characters).