Remove unused option; delay free
For hmmdmaster, in the case of hmmscant, a DNA alphabet was being
created then immediately destroyed. The result was a possibly
corrupted alphabet. This could caue the inmap to break, leading
to corrupted input sequence in the daemon master. Resolved by
delaying the free until after the sequence is parsed.