Fixes a thread data race in nhmmer.
In nhmmer, the master thread increments pli->nseqs counter, while
worker threads call p7_pli_NewSeq(), which is originally intended to
increment both the nseqs and nres counters. p7_pli_NewSeq() has an `if
( ! long_targets)` condition around the nseqs increment but it still
accesses nseqs to set pli->Z. This creates a race: the master is
accessing nseqs while workers are changing it.

The fix is to put the pli->Z update inside that `if (! long_targets)`.
This is safe (it does not affect E-value calculations), because
nhmmer/nhmmscan don't use nseqs for E-value calculations anyway (they
use nres/W), and nhmmer doesn't use lower-bound Z and E-value
calculations during the search like hmmsearch anyway; it only sets Z
and E-values at the very end, after the search completes, and it's
about to output results.

The nhmmer code should be much better organized and modularized, but
that'll have to wait for another day.

This is the third of three fixes (two in H3, one in Easel) that address
H3 issues #277 and #278.

[SRE H14/110]