Replaces all H3 sprintf() calls with snprintf
(Same changes and rationale as in Easel commit e6c56e4. A lightly
edited copy of that log message follows.)

As of macOS13/XCode14.1, clang will try to issue a compile-time
warning for _any_ use of sprintf(). This works for C++ code, but due
to (unintended?)  quirks, for C code the warnings only appear under
certain unusual conditions -- including when the Address Sanitizer
(asan) is used.

Briefly: asan sets _FORTIFY_SOURCE to 0, which keeps macOS from
undef'ing sprintf() and replacing it with a security-checking
version. Other ways to circumvent the warnings include defining
_POSIX_C_SOURCE or __cplusplus, or undefining __GNUC__.  See
https://openradar.appspot.com/FB11761475 for thorough explanation.

This commit systematically replaces all H3 calls to sprintf() with
snprintf(). We nigh-always have the buffer size n handy anyway when
we're calling sprintf(), so changing to snprintf() is easy.