Fixes an asan failure in p7_gmx_utest; adds eslENABLE_{ASAN,TSAN} #define's.
p7_gmx_utest's utest_GrowTo() failed in a Linux VM with the address
sanitizer at an extreme test that requires an 8.6G allocation to
exercise iss #176. This malloc fails, and instead of returning NULL
(which our code is designed to gracefully catch, when memory is
limited!), under asan it crashes out. asan helpfully says that we can
set allocator_may_return_null=1 in ASAN_OPTIONS to change this
behavior, but we don't easily control a user's (builder's)
environment.

Instead, the fix is to not run this particular piece of the test when
we're instrumented w/ asan or tsan.

To implement that, we need compile-time constants we can check for
whether asan or tsan are enabled. Looks like gcc and clang both
provide for this it's not standardized (i.e. on Clang you can do ifdef
__has_feature(address_sanitizer); gcc defines __SANITIZE_ADDRESS__).
Since we're enabling asan/tsan through ./configure
--enable_{asan,tsan} now, we have the config script define
eslENABLE_ASAN and eslENABLE_TSAN constants in addition to adding to
CFLAGS/LDFLAGS.

The relevant section of p7_gmx.c is now wrapped in `#if
!defined(eslENABLE_ASAN) && !defined(eslENABLE_TSAN)`.

I named them `esl` config constants because eventually we'll want them in
Easel, not just HMMER; but I haven't done that yet.